src/fase7/timer_svcbus_q_consumer/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-servicebus>=7.12.3
pymongo>=4.10.1
boto3>=1.34.69
